Improved 64-bit compatibility, allows to set data(cache) sizes larger than 2^32
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@33970 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ include/bdb/bdb_env.hpp
@@ -124,7 +124,7 @@ public:
     DB_ENV* GetEnv() { return m_Env; }
 
     /// Set cache size for the environment.
-    void SetCacheSize(unsigned int cache_size);
+    void SetCacheSize(Uint8 cache_size);
     
     /// Start transaction (DB_ENV->txn_begin)
     ///
